refactor(portal): drain World ID Actions GraphQL docs to scenes/common
Move the 9 World ID Actions operation docs (GetActionsV4, GetSingleActionV4
client+server, DeleteActionV4 client+server, GetActionV4WithApp, UpdateActionV4,
GetAppRpId, InsertActionV4) into scenes/common/world-id-actions, keeping the
codegen client/server split, so both v2 and the forthcoming v3 fork share one
copy. Pure relocation + import repoints; no behavior change.
