+++ web/scenes/Portal/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/page/index.tsx
@@ -4,7 +4,7 @@ import { ErrorPage } from "@/components/ErrorPage";
 import { SizingWrapper } from "@/components/SizingWrapper";
 import { TYPOGRAPHY, Typography } from "@/components/Typography";
 import Skeleton from "react-loading-skeleton";
-import { useGetSingleActionV4Query } from "./graphql/client/get-single-action-v4.generated";
+import { useGetSingleActionV4Query } from "@/scenes/common/world-id-actions/graphql/client/get-single-action-v4.generated";
 import { VerifiedTable } from "@/scenes/Portal/Teams/TeamId/Apps/AppId/Actions/ActionId/page/VerifiedTable";
 import { adaptNullifierV4 } from "./utils/adapt-nullifier-v4";
