+++ web/scenes/common/world-id-actions/graphql/client/delete-action-v4.graphql
