+++ web/scenes/common/world-id-actions/graphql/server/delete-action-v4.graphql
