+++ web/scenes/common/world-id-actions/graphql/server/insert-action-v4.graphql
