+++ web/scenes/common/world-id-actions/graphql/server/update-action-v4.graphql
